🔨 Refactor(#70): useGetLocationState 훅 제거, query Key 팩토리 패턴으로 수정
