+++ src/features/quiz/ui/Result.tsx
@@ -1,4 +1,4 @@
-import progressQuery from '../../../queries/usersQuery';
+import { progressQuery } from '../../../queries/usersQuery';
 import { useClientQuizStore } from '../../../store/useClientQuizStore';
 import useUserStore from '../../../store/useUserStore';
 import Quiz from '../../../types/Quiz';
