+++ src/pages/quiz/Quiz.tsx
@@ -15,7 +15,6 @@ import ShortAnswer from '../../features/quiz/ui/ShortAnswer';
 import componentMapping from '../../utils/componentMap';
 import useBeforeUnload from '../../hooks/useBeforeUnload';
 import Result from '../../features/quiz/ui/Result';
-import useGetLocationState from '../../hooks/useGetLocationState';
 import TotalResults from '../../features/quiz/ui/TotalResults';
 import isEqualArray from '../../utils/isEqualArray';
 import QuizzesQuery from '../../queries/quizzesQuery';
@@ -24,6 +23,7 @@ import usePreloadImages from '../../hooks/usePreloadImages';
 import { useEffect, useState } from 'react';
 import Header from '../../common/layout/Header';
 import useUserStore from '../../store/useUserStore';
+import { useLocation } from 'react-router-dom';
 
 //퀴즈페이지
 export default function Quiz() {
@@ -37,7 +37,7 @@ export default function Quiz() {
   //----------------------------
   const [result, setResult] = useState<boolean>(false);
   const { Modal, closeModal, openModal, isShow } = useModal();
-  const { partId } = useGetLocationState();
+  const { partId } = useLocation().state as { partId: number };
   if (!partId) {
     return <div>404</div>;
   }
