+++ src/queries/quizzesQuery.ts
@@ -1,10 +1,16 @@
 import { useQuery } from '@tanstack/react-query';
 import quizzesApis from './../apis/quizzesApis';
+const quizKeys = {
+  all: ['quizzes'],
+  filters: () => [...quizKeys.all, 'filter'],
+  parts: () => [...quizKeys.all, 'parts'] as const,
+  part: (partId: number) => [...quizKeys.parts(), partId] as const,
+};
 const QuizzesQuery = {
-  get: (params?: { sectionId?: number; partId: number }) => {
+  get: ({ partId }: { partId: number }) => {
     return useQuery({
-      queryKey: ['quizzes', params],
-      queryFn: () => quizzesApis.getquizzes(params),
+      queryKey: quizKeys.part(partId),
+      queryFn: () => quizzesApis.getQuizzes({ partId }),
     });
   },
 };
